+++ controllers/users.js
@@ -1,3 +1,5 @@
+const bcrypt = require('bcrypt');
+const jwt = require('jsonwebtoken');
 const User = require("../models/user");
 
 module.exports.get = async (req, res) => {
@@ -10,12 +12,15 @@ module.exports.get = async (req, res) => {
 };
 
 module.exports.post = async (req, res) => {
+  const saltRounds = 10;
+  const password = await bcrypt.hash(req.body.password, saltRounds);
   const user = new User({
     name: req.body.name,
+    password: password,
+    email: req.body.email,
     about: req.body.about,
     avatar: req.body.avatar
   });
-
   try {
     const newUser = await user.save();
     await res.status(201).json({data: newUser});
@@ -26,22 +31,28 @@ module.exports.post = async (req, res) => {
 
 module.exports.delete = async (req, res) => {
   try {
-    await res.user.delete();
-    res.json({message: "Successfully deleted"});
+    if (res.user._id == req.user._id) {
+      await res.user.delete();
+      res.json({message: "Successfully deleted"});
+    } else {
+      res.json({message: "Permission denied!"});
+    }
   } catch (err) {
     res.status(500).json({message: err.message});
   }
 };
 
 module.exports.update = async (req, res) => {
-  if (req.body.name != null) {
-    res.user.name = req.body.name;
+  if (res.user._id == req.user._id) {
+    if (req.body.name != null) {
+      res.user.name = req.body.name;
+    }
+    if (req.body.about != null) {
+      res.user.about = req.body.about;
+    }
+  } else {
+    res.json({message: "Permission denied!"});
   }
-
-  if (req.body.about != null) {
-    res.user.about = req.body.about;
-  }
-
   try {
     const updatedUser = await res.user.save({
       new: true,
@@ -68,3 +79,16 @@ module.exports.updateAvatar = async (req, res) => {
     res.status(400).json({message: err.message});
   }
 };
+
+module.exports.login = (req, res) => {
+  const {email, password} = req.body;
+  return User.findUserByCredentials(email, password)
+    .then((user) => {
+      const token = jwt.sign({ _id: user._id },"napoleon");
+      res.cookie('jwt',token, { httpOnly: true, maxAge: 604800 * 1000 });
+      res.send({message: "Авторизация прошла успешно!"});
+    })
+    .catch((err) => {
+      res.status(401).json({message: err.message});
+    });
+}
